+++ CHANGELOG.md
@@ -220,7 +220,7 @@
 - PURCHASE ORDER : Division by zero
 - PURCHASE ORDER : Completed Button french translation corrections and prompt added for it.
 - Purchase Order : Fix move 'Generate control invoice' button from toolbar to sidebar panel on form view.
- 
+- Invoice Line : Fix client side crash when clear the product.
 
 ## [Unreleased 5.0.7]
 ## Features